requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
SQLAlchemy>=2.0
pydantic>=2.6
jinja2>=3.1
requests>=2.31
PyYAML>=6.0
python-multipart>=0.0.9
weasyprint>=61.0 ; platform_system != 'Windows'
psycopg[binary]>=3.1
redis>=5.0
APScheduler>=3.10
PyJWT>=2.8
passlib[bcrypt]>=1.7
pytest>=7.4
rq>=1.15
